requirements.txt
python-telegram-bot==21.7
aiofiles==24.1.0
Flask==3.0.3
psycopg[binary]==3.2.3
